Add Vulkan QML backend (#467)
This PR adds native Vulkan backend using Qt's QML Vulkan RHI.
This PR breaks ABI slightly, so it's best if it targets post-Garden.

With a few global variables we could workaround it and backport it to Garden though. Note that Garden supports Vulkan via a fallback that is very slow. This PR offers native, full speed integration.

Note: Qt 5.14 is the minimum requirement, which means it's not available on Ubuntu 20.04 but rather Ubuntu 22.04. I put macros to use at least 5.15.2 because that's what I tested against.

Note: Qt 5 considers Vulkan QML interface to be 'experimental' and I've submitted a patch to Qt to improve compliance with the Vulkan standard. Ideally Gazebo should eventually migrate to Qt 6.

Signed-off-by: Matias N. Goldberg <dark_sylinc@yahoo.com.ar>
Signed-off-by: Ian Chen <ichen@osrfoundation.org>
Co-authored-by: Ian Chen <ichen@osrfoundation.org>