Use SplitView from QtQuick.Controls 2 for GzSplit
The one from QtQuick.Controls 1 has been deprecated for a while and has
been removed from Qt6. This change is a stepping stone for migrating to
Qt6.

The new SplitView is significantly different from the previous one and
so major changes had to be made. The original GzSplit component was also
designed with more capabilities in mind for the future, which
complicated the code. This has been simplified:

* Instead of determining whether to add a new split for each card added, two splits
are hard coded: (1) the main view port (2) the side panel. For every new
card added, the code simply checks adds the card to the main view port
if its empty. Otherwise, it puts it in the side panel.

* Animations have been removed. It was difficult to get consistent
behavior with different split sizes.

* Card items are returned instead of their names when calling `addCard`.
This simplifies the code that searches for items by name

* Logic to detect if cards have been collapsed manually by resizing the
card has been removed. It's not clear if it is even possible to do this
with any of the cards/plugins in Gazebo since the `minimumHeight`
constraint would prevent resizing the cards that small.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>